+++ src/App.js
@@ -1,15 +1,9 @@
 import "./App.css";
-import { Routes, Route } from "react-router-dom"
-import { Home } from "./pages/home/home"
-import { ProductsPage } from "./pages/productsPage/products";
-import Mockman from "mockman-js"
+import { AppRouter } from "./router/router"
 function App() {
   return (
     <div className="App">
-      <Routes>
-        <Route exact path = "/" element = {<Home/>} />
-        <Route path = "products" element = {<ProductsPage/>} />
-      </Routes>
+      <AppRouter />
     </div>
   );
 }
