+++ src/components/all-comp.jsx
@@ -1,0 +1,5 @@
+export { Navbar } from "./shared/navbar"
+export { Header } from "./header/header"
+export { GetCategories } from "./categories/categories"
+export { GetProducts } from "./products/product-card"
+export { Footer } from "./footer/footer"
