--- src/components/homeComp.js
@@ -1,4 +1,0 @@
-export { Navbar } from "./shared/navbar";
-export { Header } from "./header/header";
-export { Footer } from "./footer/footer";
-export { GetCategories } from "./categories/categories";
