+++ src/pages/all-pages.jsx
@@ -1,0 +1,3 @@
+export { Home } from "./home/home";
+export { ProductsPage } from "./productsPage/products";
+
